+++ fastview/views/__init__.py
@@ -1,5 +1,0 @@
-from .display import *  # noqa
-from .generics import *  # noqa
-from .inlines import *  # noqa
-from .mixins import *  # noqa
-from .objects import *  # noqa